提交预警模块40%
提交预警模块40%

+++ BiDaiBaoAPP/BiDaiBaoAPP/Class/Domain/Warning/View/BDB_TableViewCell_Two.h
@@ -11,7 +11,7 @@
 @class BDBWarningAddViewController;
 @protocol BDB_TableViewCell_TwoDelegate <NSObject>
 
-- (void)updateSliderValue:(CGFloat)sliderValue;
+- (void)updateSliderValue:(NSInteger)sliderValue;
 
 @end
 @interface BDB_TableViewCell_Two : UITableViewCell
